Some minor correction
- Remove unused comment klo ngga di pake, jadi suatu saat lo commit di project kantor nantinya ngga bikin kotor.
- Ngga perlu di convert jadi CGFloat krn return nya udh pasti CGFloat klo di ksh value berapa aja

+++ 2 - izlude/collectionView Challenge/collectionView Challenge/TableViewController.swift
@@ -64,12 +64,11 @@ extension TableViewController: UITableViewDataSource{
 extension TableViewController: UITableViewDelegate{
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let collectionVC = CollectionViewController(pallete: colors[indexPath.row])
-//        present(collectionVC, animated: true)
         navigationController?.pushViewController(collectionVC, animated: true)
         
         tableView.deselectRow(at: indexPath, animated: true)
     }
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return CGFloat(100)
+        return 100
     }
 }
